+++ projects/sample-routing-module/firmware/Raspberry_PI4_modelB/src/tx_test.c
@@ -27,8 +27,8 @@ int main(void)
     size_t tx_len = 0U;
     ssize_t rc;
     uint8_t byte;
-    int count = 0;
-    int total_len;
+    size_t count = 0U;
+    size_t total_len;
     uint8_t payload_len;
     uint16_t crc_rx;
     uint16_t crc_calc;
@@ -102,14 +102,14 @@ int main(void)
         rc = serial_port_read_byte_timeout(fd, &byte, 3000);
         if (rc < 0)
         {
-            printf("ERROR: read failed at header byte %d\n", count);
+            printf("ERROR: read failed at header byte %zu\n", count);
             serial_port_close(fd);
             return -1;
         }
 
         if (rc == 0)
         {
-            printf("ERROR: timeout waiting for header byte %d\n", count);
+            printf("ERROR: timeout waiting for header byte %zu\n", count);
             serial_port_close(fd);
             return -1;
         }
@@ -132,14 +132,14 @@ int main(void)
         rc = serial_port_read_byte_timeout(fd, &byte, 3000);
         if (rc < 0)
         {
-            printf("ERROR: read failed at byte %d\n", count);
+            printf("ERROR: read failed at byte %zu\n", count);
             serial_port_close(fd);
             return -1;
         }
 
         if (rc == 0)
         {
-            printf("ERROR: timeout waiting for byte %d\n", count);
+            printf("ERROR: timeout waiting for byte %zu\n", count);
             serial_port_close(fd);
             return -1;
         }
